Add files via upload
"Thrilled to Share My Achievement! 🎉

I'm excited to announce that I've successfully completed the certification requirements for Certified Technical Architect (CTA) from ServiceNow! 💻

This certification is a testament to my hard work and dedication to staying up-to-date with the latest technologies and best practices in the industry. I'm grateful for the opportunity to enhance my skills and knowledge, and I'm looking forward to applying them to drive success in my future endeavors.

Special thanks to ServiceNow for this amazing opportunity! 🙏

#ServiceNow #CertifiedTechnicalArchitect #CTA